LPS-147656 Move CachedLoggerProvider to new module
